VideoPress: make Library the default dashboard screen (JETPACK-1887)
Product decision from the July 8th Product Compass call: land on the
Library — the stats Overview is empty and unwelcoming for new users —
with the tab strip becoming Library / Stats / Settings.

The Library route takes over the index path and the stats screen moves
to /stats with its tab renamed to Stats. The legacy hash-router
reconciler moves with the index path (legacy links land on whichever
route owns it), and every internal navigation target follows: the video
details breadcrumb and post-delete navigation return to the Library at
/, the Stats cards' see-all links point at /, and the details-page
origin tag becomes from: 'stats' so Stats-originated visits still
return to /stats.

The routes/overview and components/overview directories keep their
names — wp-build routes by package.json route.path, not directory name
— to keep the diff reviewable.
