+++ projects/packages/videopress/changelog/update-videopress-library-default-tab
@@ -1,0 +1,4 @@
+Significance: minor
+Type: changed
+
+VideoPress: make Library the default dashboard screen and move stats to a Stats tab
