+++ projects/packages/videopress/src/dashboard/components/overview/most-viewed-card.tsx
@@ -40,7 +40,7 @@ export default function MostViewedCard( { videos, isLoading }: Props ): ReactEle
 			emptyMessage={ __( 'No videos viewed in the chosen period.', 'jetpack-videopress-pkg' ) }
 			footer={
 				<Stack direction="row" justify="center" className="vp-overview__ranking-footer">
-					<Link to="/library">{ __( 'See all videos', 'jetpack-videopress-pkg' ) }</Link>
+					<Link to="/">{ __( 'See all videos', 'jetpack-videopress-pkg' ) }</Link>
 				</Stack>
 			}
 		/>
