+++ projects/packages/videopress/src/dashboard/components/overview/top-by-watch-time-card.tsx
@@ -44,7 +44,7 @@ export default function TopByWatchTimeCard( { videos, isLoading }: Props ): Reac
 			formatValue={ formatWatchTime }
 			footer={
 				<Stack direction="row" justify="center" className="vp-overview__ranking-footer">
-					<Link to="/library">{ __( 'See all videos', 'jetpack-videopress-pkg' ) }</Link>
+					<Link to="/">{ __( 'See all videos', 'jetpack-videopress-pkg' ) }</Link>
 				</Stack>
 			}
 		/>
